requirements.txt
py-dateutil
docker-py>=1.0.0
